requirements.txt
streamlit
google-play-scraper
anthropic
python-dotenv
